@@ -1,6 +1,6 @@
 {
   "name": "altinn-app-frontend",
-  "version": "3.6.5",
+  "version": "3.6.6",
   "description": "",
   "main": "index.js",
   "scripts": {
@@ -36,10 +36,10 @@
     "jsonpointer": "4.1.0",
     "lodash.isequal": "~4.5.0",
     "moment": "~2.29.1",
-    "react": "~16.13.1",
+    "react": "~17.0.2",
     "react-content-loader": "~6.0.3",
     "react-device-detect": "~1.17.0",
-    "react-dom": "~16.13.1",
+    "react-dom": "~17.0.2",
     "react-dropzone": "~11.3.2",
     "react-number-format": "4.6.0",
     "react-redux": "~7.2.4",
@@ -59,11 +59,11 @@
     "@types/enzyme": "~3.10.8",
     "@types/jest": "~26.0.23",
     "@types/lodash.isequal": "~4.5.5",
-    "@types/react": "~16.9.56",
-    "@types/react-dom": "~16.9.12",
+    "@types/react": "~17.0.9",
+    "@types/react-dom": "~17.0.6",
     "@types/react-redux": "~7.1.16",
     "@types/react-router-dom": "~5.1.7",
-    "@types/react-test-renderer": "~16.9.5",
+    "@types/react-test-renderer": "~17.0.1",
     "@types/redux-mock-store": "~1.0.2",
     "@types/uuid": "8.3.0",
     "altinn-designsystem": "~3.0.3",
@@ -72,7 +72,7 @@
     "cross-env": "~7.0.3",
     "css-loader": "~5.2.6",
     "enzyme": "~3.11.0",
-    "enzyme-adapter-react-16": "~1.15.6",
+    "@wojtekmaj/enzyme-adapter-react-17": "~0.6.1",
     "eslint": "~7.27.0",
     "eslint-config-airbnb-base": "~14.2.1",
     "fork-ts-checker-webpack-plugin": "~6.2.10",
@@ -80,11 +80,11 @@
     "git-revision-webpack-plugin": "~3.0.6",
     "html-loader": "~1.3.2",
     "html-webpack-plugin": "~4.5.2",
-    "jest": "~26.6.3",
+    "jest": "~27.0.4",
     "jest-junit": "~12.1.0",
     "mini-css-extract-plugin": "~1.6.0",
     "react-styleguidist": "~11.1.7",
-    "react-test-renderer": "~16.13.1",
+    "react-test-renderer": "~17.0.2",
     "redux-mock-store": "~1.5.4",
     "redux-saga-test-plan": "4.0.1",
     "rimraf": "~3.0.2",
@@ -92,7 +92,7 @@
     "style-loader": "~1.3.0",
     "svg-inline-loader": "~0.8.2",
     "terser-webpack-plugin": "~4.2.3",
-    "ts-jest": "~26.5.6",
+    "ts-jest": "~27.0.3",
     "ts-loader": "~8.1.0",
     "webpack": "~4.44.2",
     "webpack-cli": "~3.3.12",
@@ -129,6 +129,7 @@
     ],
     "snapshotSerializers": [
       "enzyme-to-json/serializer"
-    ]
+    ],
+    "testEnvironment": "jsdom"
   }
 }
