@@ -23,28 +23,28 @@
     "@material-ui/core": "~4.11.4",
     "axios": "~0.21.1",
     "moment": "~2.29.1",
-    "react": "~16.13.1",
-    "react-dom": "~16.13.1"
+    "react": "~17.0.2",
+    "react-dom": "~17.0.2"
   },
   "devDependencies": {
     "@babel/core": "~7.14.3",
     "@babel/preset-env": "~7.14.4",
     "@babel/preset-react": "~7.13.13",
     "@types/jest": "~26.0.23",
-    "@types/react": "~16.9.56",
-    "@types/react-dom": "~16.9.12",
+    "@types/react": "~17.0.9",
+    "@types/react-dom": "~17.0.6",
     "cross-env": "~7.0.3",
     "css-loader": "~5.2.6",
     "enzyme": "~3.11.0",
-    "enzyme-adapter-react-16": "~1.15.6",
+    "@wojtekmaj/enzyme-adapter-react-17": "~0.6.1",
     "fork-ts-checker-webpack-plugin": "~6.2.10",
     "fork-ts-checker-notifier-webpack-plugin": "3.0.0",
     "html-webpack-plugin": "~4.5.2",
-    "jest": "~26.6.3",
+    "jest": "~27.0.4",
     "mini-css-extract-plugin": "~1.6.0",
     "rimraf": "~3.0.2",
     "terser-webpack-plugin": "~4.2.3",
-    "ts-jest": "~26.5.6",
+    "ts-jest": "~27.0.3",
     "ts-loader": "~8.1.0",
     "typescript": "~4.3.2",
     "webpack": "~4.44.2",
@@ -67,7 +67,8 @@
       "js"
     ],
     "setupTestFrameworkScriptFile": "<rootDir>/setupTests.js",
-    "testURL": "http://localhost/receipt/mockInstanceOwnerId/6697de17-18c7-4fb9-a428-d6a414a797ae"
+    "testURL": "http://localhost/receipt/mockInstanceOwnerId/6697de17-18c7-4fb9-a428-d6a414a797ae",
+    "testEnvironment": "jsdom"
   },
   "browserslist": [
     ">0.2%",
